add cast

+++ src/secsse_loglik.cpp
@@ -152,8 +152,8 @@ inte_nodes_t find_inte_nodes(std::vector<std::vector<double>>& phy_edge, const s
     if ((*it0)[1] > (*it1)[1]) {
       std::swap(*it0, *it1);
     }
-    inode.desc[0] = { &(*states)[(*it0)[1] - 1], (*it0)[2], (*it0)[1] };
-    inode.desc[1] = { &(*states)[(*it1)[1] - 1], (*it1)[2], (*it1)[1] };
+    inode.desc[0] = { &(*states)[(*it0)[1] - 1], (*it0)[2], static_cast<int>((*it0)[1]) };
+    inode.desc[1] = { &(*states)[(*it1)[1] - 1], (*it1)[2], static_cast<int>((*it1)[1]) };
   };
   return res;
 }
